+++ DatabaseManager.py
@@ -1,0 +1,77 @@
+
+from constants import *
+from menu_definitions import add_menu, delete_menu, list_menu
+from db_connection import Session
+# Note that until you import your SQLAlchemy declarative classes, such as Student, Python
+# will not execute that code, and SQLAlchemy will be unaware of the mapped table.
+from Department import Department
+from Major import Major
+from Student import Student
+from Option import Option
+from Menu import Menu
+
+from pprint import pprint
+
+from QueryAdd import add_major
+
+
+
+def add(sess: Session):
+    add_action: str = ''
+    while add_action != add_menu.last_action():
+        add_action = add_menu.menu_prompt()
+        exec(add_action)
+
+
+def delete(sess: Session):
+    delete_action: str = ''
+    while delete_action != delete_menu.last_action():
+        delete_action = delete_menu.menu_prompt()
+        exec(delete_action)
+
+
+def list_objects(sess: Session):
+    list_action: str = ''
+    while list_action != list_menu.last_action():
+        list_action = list_menu.menu_prompt()
+        exec(list_action)
+
+
+
+def boilerplate(sess: Session):
+    """
+    Add boilerplate data initially to jump start the testing.  Remember that there is no
+    checking of this data, so only run this option once from the console, or you will
+    get a uniqueness constraint violation from the database.
+    :param sess:    The session that's open.
+    :return:        None
+    """
+    department: Department = Department('CECS', 'Computer Engineering Computer Science')
+    major1: Major = Major(department, 'Computer Science', 'Fun with blinking lights')
+    major2: Major = Major(department, 'Computer Engineering', 'Much closer to the silicon')
+    student1: Student = Student('Brown', 'David', 'david.brown@gmail.com')
+    student2: Student = Student('Brown', 'Mary', 'marydenni.brown@gmail.com')
+    student3: Student = Student('Disposable', 'Bandit', 'disposable.bandit@gmail.com')
+    student1.add_major(major1)
+    student2.add_major(major1)
+    student2.add_major(major2)
+    sess.add(department)
+    sess.add(major1)
+    sess.add(major2)
+    sess.add(student1)
+    sess.add(student2)
+    sess.add(student3)
+    sess.flush()                                # Force SQLAlchemy to update the database, although not commit
+
+
+def session_rollback(sess):
+    """
+    Give the user a chance to roll back to the most recent commit point.
+    :param sess:    The connection to the database.
+    :return:        None
+    """
+    confirm_menu = Menu('main', 'Please select one of the following options:', [
+        Option("Yes, I really want to roll back this session", "sess.rollback()"),
+        Option("No, I hit this option by mistake", "pass")
+    ])
+    exec(confirm_menu.menu_prompt())
