+++ Department.py
@@ -6,26 +6,18 @@ from typing import List                 # Use this for the list of courses offer
 
 
 class Department(Base):
-    """An organization within a particular college within a university.  Each
-    department offers one or more major fields of study to its students, and
-    within each major, some number of courses.  Each course is offered on
-    a regular basis as a scheduled section of a given course.
-
-    Note, this is just a shell of the Department class.  There are additional
-    columns needed, but this is enough to demonstrate one-to-many relationships."""
     __tablename__ = "departments"  # Give SQLAlchemy th name of the table.
-    abbreviation: Mapped[str] = mapped_column('abbreviation', String,
-                                              nullable=False, primary_key=True)
+    # Primary key
+    abbreviation: Mapped[str] = mapped_column('abbreviation', String, nullable=False, primary_key=True)
+    
     name: Mapped[str] = mapped_column('name', String(50), nullable=False)
-    # The list of majors in this department
+    # Relationships
     majors: Mapped[List["Major"]] = relationship(back_populates="department")
-    # The list of courses offered by this department
     courses: Mapped[List["Course"]] = relationship(back_populates="department")
-    # __table_args__ can best be viewed as directives that we ask SQLAlchemy to
-    # send to the database.  In this case, that we want two separate uniqueness
-    # constraints (candidate keys).
+    # Unniqueness constraints 
     __table_args__ = (UniqueConstraint("name", name="departments_uk_01"), )
 
+    # "Constructor"
     def __init__(self, abbreviation: str, name: str):
         self.abbreviation = abbreviation
         self.name = name
