--- QueryCourse.py
@@ -1,129 +1,0 @@
-from Course import Course 
-
-def add_course(session: Session):
-    """
-    This demonstrates how to use the utilities in SQLAlchemy Utilities for checking
-    all the uniqueness constraints of a table in one method call.  The user
-    experience is tougher to customize using this technique, but the benefit is that
-    new uniqueness constraints will be automatically added to the list to be checked,
-    without any change to the add_course code.
-
-    Prompt the user for the information for a new course and validate
-    the input to make sure that we do not create any duplicates.
-    :param session: The connection to the database.
-    :return:        None
-    """
-    print("Which department offers this course?")
-    department: Department = select_department(sess)
-    description: str = input('Please enter the course description-->')
-    units: int = int(input('How many units for this course-->'))
-    violation = True  # Flag that we still have to prompt for fresh values
-    while violation:
-        name = input("Course full name--> ")
-        number = int(input("Course number--> "))
-        course = Course(department, number, name, description, units)
-        violated_constraints = check_unique(Session, course)
-        if len(violated_constraints) > 0:
-            print('The following uniqueness constraints were violated:')
-            pprint(violated_constraints)
-            print('Please try again.')
-        else:
-            violation = False
-    session.add(course)
-
-
-def select_course(sess: Session) -> Course:
-    """
-    Select a course by the combination of the department abbreviation and course number.
-    Note, a similar query would be to select the course on the basis of the department
-    abbreviation and the course name.
-    :param sess:    The connection to the database.
-    :return:        The selected student.
-    """
-    found: bool = False
-    department_abbreviation: str = ''
-    course_number: int = -1
-    while not found:
-        department_abbreviation = input("Department abbreviation--> ")
-        course_number = int(input("Course Number--> "))
-        name_count: int = sess.query(Course).filter(Course.departmentAbbreviation == department_abbreviation,
-                                                    Course.courseNumber == course_number).count()
-        found = name_count == 1
-        if not found:
-            print("No course by that number in that department.  Try again.")
-    course = sess.query(Course).filter(Course.departmentAbbreviation == department_abbreviation,
-                                       Course.courseNumber == course_number).first()
-    return course
-
-
-def delete_course(session):
-    """
-    Prompt the user to select a course by department abbreviation and course number, then delete it.
-    :param session: The connection to the database.
-    :return: None
-    """
-    print("Deleting a course")
-    course = select_course(session)
-    if 0 < session.query(Sections).filter(Section.courseNumber == course.courseNumber).count():
-        print("Sections depend on this course, go delete them first and try again.")
-    else:
-        session.delete(course)
-
-
-
-def list_course(sess: Session):
-    """
-    List all courses currently in the database.
-    :param sess:    The connection to the database.
-    :return:        None
-    """
-    # session.query returns an iterator.  The list function converts that iterator
-    # into a list of elements.  In this case, they are instances of the Student class.
-    courses: [Course] = list(sess.query(Course).order_by(Course.courseNumber))
-    for course in courses:
-        print(course)
-
-
-def move_course_to_new_department(sess: Session):
-    """
-    Take an existing course and move it to an existing department.  The course has to
-    have a department when the course is created, so this routine just moves it from
-    one department to another.
-
-    The change in department has to occur from the Course end of the association because
-    the association is mandatory.  We cannot have the course not have any department for
-    any time the way that we would if we moved it to a new department from the department
-    end.
-
-    Also, the change in department requires that we make sure that the course will not
-    conflict with any existing courses in the new department by name or number.
-    :param sess:    The connection to the database.
-    :return:        None
-    """
-    print("Input the course to move to a new department.")
-    course = select_course(sess)
-    old_department = course.department
-    print("Input the department to move that course to.")
-    new_department = select_department(sess)
-    if new_department == old_department:
-        print("Error, you're not moving to a different department.")
-    else:
-        # check to be sure that we are not violating the {departmentAbbreviation, name} UK.
-        name_count: int = sess.query(Course).filter(Course.departmentAbbreviation == new_department.abbreviation,
-                                                    Course.name == course.name).count()
-        unique_name = name_count == 0
-        if not unique_name:
-            print("We already have a course by that name in that department.  Try again.")
-        if unique_name:
-            # Make sure that moving the course will not violate the {departmentAbbreviation,
-            # course number} uniqueness constraint.
-            number_count = sess.query(Course). \
-                filter(Course.departmentAbbreviation == new_department.abbreviation,
-                       Course.courseNumber == course.courseNumber).count()
-            if number_count != 0:
-                print("We already have a course by that number in that department.  Try again.")
-            else:
-                course.set_department(new_department)
-
-
-
