+++ QueryList.py
@@ -1,0 +1,101 @@
+from Department import Department
+from Course import Course
+from Student import Student
+from Major import Major
+from StudentMajor import StudentMajor
+from QuerySelect import select_department
+from db_connection import Session
+
+
+def list_department(session: Session):
+    """
+    List all departments, sorted by the abbreviation.
+    :param session:     The connection to the database.
+    :return:            None
+    """
+    # session.query returns an iterator.  The list function converts that iterator
+    # into a list of elements.  In this case, they are instances of the Student class.
+    departments: [Department] = list(session.query(Department).order_by(Department.abbreviation))
+    for department in departments:
+        print(department)
+
+
+def list_course(sess: Session):
+    """
+    List all courses currently in the database.
+    :param sess:    The connection to the database.
+    :return:        None
+    """
+    # session.query returns an iterator.  The list function converts that iterator
+    # into a list of elements.  In this case, they are instances of the Student class.
+    courses: [Course] = list(sess.query(Course).order_by(Course.courseNumber))
+    for course in courses:
+        print(course)
+
+
+def list_department_courses(sess):
+    department = select_department(sess)
+    dept_courses: [Course] = department.get_courses()
+    print("Course for department: " + str(department))
+    for dept_course in dept_courses:
+        print(dept_course)
+
+
+def list_student(sess: Session):
+    """
+    List all Students currently in the database.
+    :param sess:    The current connection to the database.
+    :return:
+    """
+    students: [Student] = list(sess.query(Student).order_by(Student.lastName, Student.firstName))
+    for student in students:
+        print(student)
+
+
+def list_student_major(sess: Session):
+    """Prompt the user for the student, and then list the majors that the student has declared.
+    :param sess:    The connection to the database
+    :return:        None
+    """
+    student: Student = select_student(sess)
+    recs = sess.query(Student).join(StudentMajor, Student.studentID == StudentMajor.studentId).join(
+        Major, StudentMajor.majorName == Major.name).filter(
+        Student.studentID == student.studentID).add_columns(
+        Student.lastName, Student.firstName, Major.description, Major.name).all()
+    for stu in recs:
+        print(f"Student name: {stu.lastName}, {stu.firstName}, Major: {stu.name}, Description: {stu.description}")
+
+
+def list_major(sess: Session):
+    """
+    List all majors in the database.
+    :param sess:    The current connection to the database.
+    :return:
+    """
+    majors: [Major] = list(sess.query(Major).order_by(Major.departmentAbbreviation))
+    for major in majors:
+        print(major)
+
+
+def list_major_student(sess: Session):
+    """Prompt the user for the major, then list the students who have that major declared.
+    :param sess:    The connection to the database.
+    :return:        None
+    """
+    major: Major = select_major(sess)
+    recs = sess.query(Major).join(StudentMajor, StudentMajor.majorName == Major.name).join(
+        Student, StudentMajor.studentId == Student.studentID).filter(
+        Major.name == major.name).add_columns(
+        Student.lastName, Student.firstName, Major.description, Major.name).all()
+    for stu in recs:
+        print(f"Student name: {stu.lastName}, {stu.firstName}, Major: {stu.name}, Description: {stu.description}")
+
+
+
+
+
+
+
+
+
+
