+++ QuerySelect.py
@@ -1,0 +1,193 @@
+from Section import Section
+from Department import Department
+from Course import Course
+from Student import Student
+from Major import Major
+from db_connection import Session
+
+
+def select_department(session: Session) -> Department:
+    """
+    Prompt the user for a specific department by the department abbreviation.
+    :param sess:    The connection to the database.
+    :return:        The selected department.
+    """
+    found: bool = False
+    abbreviation: str = ''
+    while not found:
+        abbreviation = input("Enter the department abbreviation--> ")
+        abbreviation_count: int = session.query(Department). \
+            filter(Department.abbreviation == abbreviation).count()
+        found = abbreviation_count == 1
+        if not found:
+            print("No department with that abbreviation.  Try again.")
+    return_department: Department = sess.query(Department). \
+        filter(Department.abbreviation == abbreviation).first()
+    return return_department
+
+
+def select_course(sess: Session) -> Course:
+    """
+    Select a course by the combination of the department abbreviation and course number.
+    Note, a similar query would be to select the course on the basis of the department
+    abbreviation and the course name.
+    :param sess:    The connection to the database.
+    :return:        The selected student.
+    """
+    found: bool = False
+    department_abbreviation: str = ''
+    course_number: int = -1
+    while not found:
+        department_abbreviation = input("Department abbreviation--> ")
+        course_number = int(input("Course Number--> "))
+        name_count: int = sess.query(Course).filter(Course.departmentAbbreviation == department_abbreviation,
+                                                    Course.courseNumber == course_number).count()
+        found = name_count == 1
+        if not found:
+            print("No course by that number in that department.  Try again.")
+    course = sess.query(Course).filter(Course.departmentAbbreviation == department_abbreviation,
+                                       Course.courseNumber == course_number).first()
+    return course
+
+
+def select_section(sess: Session):
+    """
+    Select a section by its attributes.
+    :param sess: The connection to the database.
+    :return: The selected section.
+    """
+    found = False
+    while not found:
+        department_abbreviation = input("Department Abbreviation --> ")
+        course_number = int(input("Course Number --> "))
+        section_number = int(input("Section Number --> "))
+        semester = input("Semester --> ")
+        section_year = int(input("Section Year --> "))
+
+        count = sess.query(Section).filter(Section.departmentAbbreviation == department_abbreviation,
+                   Section.courseNumber == course_number,
+                   Section.sectionNumber == section_number,
+                   Section.semester == semester,
+                   Section.sectionYear == section_year).count()
+        found = count == 1
+        if not found:
+          print("No section found with that information. Please try again.")
+    # Otherwise its been found
+    return sess.query(Section).filter(Section.departmentAbbreviation == department_abbreviation,
+            Section.courseNumber == course_number, Section.sectionNumber == section_number,
+           Section.semester == semester, Section.sectionYear == section_year).first()
+
+def move_course_to_new_department(sess: Session):
+    """
+    Take an existing course and move it to an existing department.  The course has to
+    have a department when the course is created, so this routine just moves it from
+    one department to another.
+
+    The change in department has to occur from the Course end of the association because
+    the association is mandatory.  We cannot have the course not have any department for
+    any time the way that we would if we moved it to a new department from the department
+    end.
+
+    Also, the change in department requires that we make sure that the course will not
+    conflict with any existing courses in the new department by name or number.
+    :param sess:    The connection to the database.
+    :return:        None
+    """
+    print("Input the course to move to a new department.")
+    course = select_course(sess)
+    old_department = course.department
+    print("Input the department to move that course to.")
+    new_department = select_department(sess)
+    if new_department == old_department:
+        print("Error, you're not moving to a different department.")
+    else:
+        # check to be sure that we are not violating the {departmentAbbreviation, name} UK.
+        name_count: int = sess.query(Course).filter(Course.departmentAbbreviation == new_department.abbreviation,
+                                                    Course.name == course.name).count()
+        unique_name = name_count == 0
+        if not unique_name:
+            print("We already have a course by that name in that department.  Try again.")
+        if unique_name:
+            # Make sure that moving the course will not violate the {departmentAbbreviation,
+            # course number} uniqueness constraint.
+            number_count = sess.query(Course). \
+                filter(Course.departmentAbbreviation == new_department.abbreviation,
+                       Course.courseNumber == course.courseNumber).count()
+            if number_count != 0:
+                print("We already have a course by that number in that department.  Try again.")
+            else:
+                course.set_department(new_department)
+
+
+def select_student(sess: Session) -> Student:
+    """
+    Select a student by the combination of the last and first.
+    :param sess:    The connection to the database.
+    :return:        The selected student.
+    """
+    found: bool = False
+    last_name: str = ''
+    first_name: str = ''
+    while not found:
+        last_name = input("Student's last name--> ")
+        first_name = input("Student's first name--> ")
+        name_count: int = sess.query(Student).filter(Student.lastName == last_name,
+                                                     Student.firstName == first_name).count()
+        found = name_count == 1
+        if not found:
+            print("No student found by that name.  Try again.")
+    student: Student = sess.query(Student).filter(Student.lastName == last_name,
+                                                  Student.firstName == first_name).first()
+    return student
+
+
+def select_major(sess: Session) -> Major:
+    """
+    Select a major by its name.
+    :param sess:    The connection to the database.
+    :return:        The selected student.
+    """
+    found: bool = False
+    name: str = ''
+    while not found:
+        name = input("Major's name--> ")
+        name_count: int = sess.query(Major).filter(Major.name == name).count()
+        found = name_count == 1
+        if not found:
+            print("No major found by that name.  Try again.")
+    major: Major = sess.query(Major).filter(Major.name == name).first()
+    return major
+
+
+def select_student_from_list(session: Session):
+    """
+    This is just a cute little use of the Menu object.  Basically, I create a
+    menu on the fly from data selected from the database, and then use the
+    menu_prompt method on Menu to display characteristic descriptive data, with
+    an index printed out with each entry, and prompt the user until they select
+    one of the Students.
+    :param session:     The connection to the database.
+    :return:            None
+    """
+    # query returns an iterator of Student objects, I want to put those into a list.  Technically,
+    # that was not necessary, I could have just iterated through the query output directly.
+    students: [Department] = list(sess.query(Department).order_by(Department.lastName, Department.firstName))
+    options: [Option] = []  # The list of menu options that we're constructing.
+    for student in students:
+        # Each time we construct an Option instance, we put the full name of the student into
+        # the "prompt" and then the student ID (albeit as a string) in as the "action".
+        options.append(Option(student.lastName + ', ' + student.firstName, student.studentId))
+    temp_menu = Menu('Student list', 'Select a student from this list', options)
+    # text_studentId is the "action" corresponding to the student that the user selected.
+    text_studentId: str = temp_menu.menu_prompt()
+    # get that student by selecting based on the int version of the student id corresponding
+    # to the student that the user selected.
+    returned_student = sess.query(Department).filter(Department.studentId == int(text_studentId)).first()
+    # this is really just to prove the point.  Ideally, we would return the student, but that
+    # will present challenges in the exec call, so I didn't bother.
+    print("Selected student: ", returned_student)
+
+
+
+
+
