--- QueryStudent.py
@@ -1,131 +1,0 @@
-from Student import Student
-from Department import Department
-'''
-    Basically anything that starts with looking up a specific student 
-    is found here.
-'''
-
-
-def add_student(session: Session):
-    """
-    Prompt the user for the information for a new student and validate
-    the input to make sure that we do not create any duplicates.
-    :param session: The connection to the database.
-    :return:        None
-    """
-    unique_name: bool = False
-    unique_email: bool = False
-    last_name: str = ''
-    first_name: str = ''
-    email: str = ''
-    while not unique_email or not unique_name:
-        last_name = input("Student last name--> ")
-        first_name = input("Student first name-->")
-        email = input("Student e-mail address--> ")
-        name_count: int = session.query(Student).filter(Student.lastName == last_name,
-                                                        Student.firstName == first_name).count()
-        unique_name = name_count == 0
-        if not unique_name:
-            print("We already have a student by that name.  Try again.")
-        if unique_name:
-            email_count = session.query(Student).filter(Student.email == email).count()
-            unique_email = email_count == 0
-            if not unique_email:
-                print("We already have a student with that email address.  Try again.")
-    new_student = Student(last_name, first_name, email)
-    session.add(new_student)
-
-
-def select_student(sess) -> Student:
-    """
-    Select a student by the combination of the last and first.
-    :param sess:    The connection to the database.
-    :return:        The selected student.
-    """
-    found: bool = False
-    last_name: str = ''
-    first_name: str = ''
-    while not found:
-        last_name = input("Student's last name--> ")
-        first_name = input("Student's first name--> ")
-        name_count: int = sess.query(Student).filter(Student.lastName == last_name,
-                                                     Student.firstName == first_name).count()
-        found = name_count == 1
-        if not found:
-            print("No student found by that name.  Try again.")
-    student: Student = sess.query(Student).filter(Student.lastName == last_name,
-                                                  Student.firstName == first_name).first()
-    return student
-
-def delete_student(session: Session):
-    """
-    Prompt the user for a student to delete and delete them.
-    :param session:     The current connection to the database.
-    :return:            None
-    """
-    student: Student = select_student(session)
-    """This is a bit ghetto.  The relationship from Student to StudentMajor has 
-    cascade delete, so this delete will work even if a student has declared one
-    or more majors.  I could write a method on Student that would return some
-    indication of whether it has any children, and use that to let the user know
-    that they cannot delete this particular student.  But I'm too lazy at this
-    point.
-    """
-    session.delete(student)
-
-
-def list_student(sess: Session):
-    """
-    List all Students currently in the database.
-    :param sess:    The current connection to the database.
-    :return:
-    """
-    students: [Student] = list(sess.query(Student).order_by(Student.lastName, Student.firstName))
-    for student in students:
-        print(student)
-
-
-def list_student_major(sess: Session):
-    """Prompt the user for the student, and then list the majors that the student has declared.
-    :param sess:    The connection to the database
-    :return:        None
-    """
-    student: Student = select_student(sess)
-    recs = sess.query(Student).join(StudentMajor, Student.studentID == StudentMajor.studentId).join(
-        Major, StudentMajor.majorName == Major.name).filter(
-        Student.studentID == student.studentID).add_columns(
-        Student.lastName, Student.firstName, Major.description, Major.name).all()
-    for stu in recs:
-        print(f"Student name: {stu.lastName}, {stu.firstName}, Major: {stu.name}, Description: {stu.description}")
-
-
-def select_student_from_list(session):
-    """
-    This is just a cute little use of the Menu object.  Basically, I create a
-    menu on the fly from data selected from the database, and then use the
-    menu_prompt method on Menu to display characteristic descriptive data, with
-    an index printed out with each entry, and prompt the user until they select
-    one of the Students.
-    :param session:     The connection to the database.
-    :return:            None
-    """
-    # query returns an iterator of Student objects, I want to put those into a list.  Technically,
-    # that was not necessary, I could have just iterated through the query output directly.
-    students: [Department] = list(sess.query(Department).order_by(Department.lastName, Department.firstName))
-    options: [Option] = []  # The list of menu options that we're constructing.
-    for student in students:
-        # Each time we construct an Option instance, we put the full name of the student into
-        # the "prompt" and then the student ID (albeit as a string) in as the "action".
-        options.append(Option(student.lastName + ', ' + student.firstName, student.studentId))
-    temp_menu = Menu('Student list', 'Select a student from this list', options)
-    # text_studentId is the "action" corresponding to the student that the user selected.
-    text_studentId: str = temp_menu.menu_prompt()
-    # get that student by selecting based on the int version of the student id corresponding
-    # to the student that the user selected.
-    returned_student = sess.query(Department).filter(Department.studentId == int(text_studentId)).first()
-    # this is really just to prove the point.  Ideally, we would return the student, but that
-    # will present challenges in the exec call, so I didn't bother.
-    print("Selected student: ", returned_student)
-
-
-
