+++ main.py
@@ -7,357 +7,19 @@ from orm_base import metadata
 # Note that until you import your SQLAlchemy declarative classes, such as Student, Python
 # will not execute that code, and SQLAlchemy will be unaware of the mapped table.
 from Department import Department
-from Course import Course
-from Section import Section
 from Major import Major
 from Student import Student
 from StudentMajor import StudentMajor
 from Option import Option
 from Menu import Menu
-from SQLAlchemyUtilities import check_unique
 from pprint import pprint
 
+from QueryAdd import add_department, add_course, add_major, add_section, add_major_student, add_student, add_student_major
+from QuerySelect import select_course, select_department, select_major, select_section, select_student, select_student_from_list
+from QueryDelete import delete_course, delete_department, delete_major_student, delete_section, delete_student, delete_student_major
+from QueryList import list_course, list_department, list_department_courses, list_major, list_major_student, list_student, list_student_major
 
-def add(sess: Session):
-    add_action: str = ''
-    while add_action != add_menu.last_action():
-        add_action = add_menu.menu_prompt()
-        exec(add_action)
-
-
-def delete(sess: Session):
-    delete_action: str = ''
-    while delete_action != delete_menu.last_action():
-        delete_action = delete_menu.menu_prompt()
-        exec(delete_action)
-
-
-def list_objects(sess: Session):
-    list_action: str = ''
-    while list_action != list_menu.last_action():
-        list_action = list_menu.menu_prompt()
-        exec(list_action)
-
-
-
-
-
-
-
-
-
-        
-
-
-def add_major(session: Session):
-    """
-    Prompt the user for the information for a new major and validate
-    the input to make sure that we do not create any duplicates.
-    :param session: The connection to the database.
-    :return:        None
-    """
-    print("Which department offers this major?")
-    department: Department = select_department(sess)
-    unique_name: bool = False
-    name: str = ''
-    while not unique_name:
-        name = input("Major name--> ")
-        name_count: int = session.query(Major).filter(Major.departmentAbbreviation == department.abbreviation,
-                                                      ).count()
-        unique_name = name_count == 0
-        if not unique_name:
-            print("We already have a major by that name in that department.  Try again.")
-    description: str = input('Please give this major a description -->')
-    major: Major = Major(department, name, description)
-    session.add(major)
-
-
-def add_student(session: Session):
-    """
-    Prompt the user for the information for a new student and validate
-    the input to make sure that we do not create any duplicates.
-    :param session: The connection to the database.
-    :return:        None
-    """
-    unique_name: bool = False
-    unique_email: bool = False
-    last_name: str = ''
-    first_name: str = ''
-    email: str = ''
-    while not unique_email or not unique_name:
-        last_name = input("Student last name--> ")
-        first_name = input("Student first name-->")
-        email = input("Student e-mail address--> ")
-        name_count: int = session.query(Student).filter(Student.lastName == last_name,
-                                                        Student.firstName == first_name).count()
-        unique_name = name_count == 0
-        if not unique_name:
-            print("We already have a student by that name.  Try again.")
-        if unique_name:
-            email_count = session.query(Student).filter(Student.email == email).count()
-            unique_email = email_count == 0
-            if not unique_email:
-                print("We already have a student with that email address.  Try again.")
-    new_student = Student(last_name, first_name, email)
-    session.add(new_student)
-
-
-def add_student_major(sess):
-    student: Student = select_student(sess)
-    major: Major = select_major(sess)
-    student_major_count: int = sess.query(StudentMajor).filter(StudentMajor.studentId == student.studentID,
-                                                               StudentMajor.majorName == major.name).count()
-    unique_student_major: bool = student_major_count == 0
-    while not unique_student_major:
-        print("That student already has that major.  Try again.")
-        student = select_student(sess)
-        major = select_major(sess)
-    student.add_major(major)
-    """The student object instance is mapped to a specific row in the Student table.  But adding
-    the new major to its list of majors does not add the new StudentMajor instance to this session.
-    That StudentMajor instance was created and added to the Student's majors list inside of the
-    add_major method, but we don't have easy access to it from here.  And I don't want to have to 
-    pass sess to the add_major method.  So instead, I add the student to the session.  You would
-    think that would cause an insert, but SQLAlchemy is smart enough to know that this student 
-    has already been inserted, so the add method takes this to be an update instead, and adds
-    the new instance of StudentMajor to the session.  THEN, when we flush the session, that 
-    transient instance of StudentMajor gets inserted into the database, and is ready to be 
-    committed later (which happens automatically when we exit the application)."""
-    sess.add(student)                           # add the StudentMajor to the session
-    sess.flush()
-
-
-def add_major_student(sess):
-    major: Major = select_major(sess)
-    student: Student = select_student(sess)
-    student_major_count: int = sess.query(StudentMajor).filter(StudentMajor.studentId == student.studentID,
-                                                               StudentMajor.majorName == major.name).count()
-    unique_student_major: bool = student_major_count == 0
-    while not unique_student_major:
-        print("That major already has that student.  Try again.")
-        major = select_major(sess)
-        student = select_student(sess)
-    major.add_student(student)
-    """The major object instance is mapped to a specific row in the Major table.  But adding
-    the new student to its list of students does not add the new StudentMajor instance to this session.
-    That StudentMajor instance was created and added to the Major's students list inside of the
-    add_student method, but we don't have easy access to it from here.  And I don't want to have to 
-    pass sess to the add_student method.  So instead, I add the major to the session.  You would
-    think that would cause an insert, but SQLAlchemy is smart enough to know that this major 
-    has already been inserted, so the add method takes this to be an update instead, and adds
-    the new instance of StudentMajor to the session.  THEN, when we flush the session, that 
-    transient instance of StudentMajor gets inserted into the database, and is ready to be 
-    committed later (which happens automatically when we exit the application)."""
-    sess.add(major)                           # add the StudentMajor to the session
-    sess.flush()
-
-
-
-
-
-
-def select_student(sess) -> Student:
-    """
-    Select a student by the combination of the last and first.
-    :param sess:    The connection to the database.
-    :return:        The selected student.
-    """
-    found: bool = False
-    last_name: str = ''
-    first_name: str = ''
-    while not found:
-        last_name = input("Student's last name--> ")
-        first_name = input("Student's first name--> ")
-        name_count: int = sess.query(Student).filter(Student.lastName == last_name,
-                                                     Student.firstName == first_name).count()
-        found = name_count == 1
-        if not found:
-            print("No student found by that name.  Try again.")
-    student: Student = sess.query(Student).filter(Student.lastName == last_name,
-                                                  Student.firstName == first_name).first()
-    return student
-
-
-def select_major(sess) -> Major:
-    """
-    Select a major by its name.
-    :param sess:    The connection to the database.
-    :return:        The selected student.
-    """
-    found: bool = False
-    name: str = ''
-    while not found:
-        name = input("Major's name--> ")
-        name_count: int = sess.query(Major).filter(Major.name == name).count()
-        found = name_count == 1
-        if not found:
-            print("No major found by that name.  Try again.")
-    major: Major = sess.query(Major).filter(Major.name == name).first()
-    return major
-
-
-
-def delete_student(session: Session):
-    """
-    Prompt the user for a student to delete and delete them.
-    :param session:     The current connection to the database.
-    :return:            None
-    """
-    student: Student = select_student(session)
-    """This is a bit ghetto.  The relationship from Student to StudentMajor has 
-    cascade delete, so this delete will work even if a student has declared one
-    or more majors.  I could write a method on Student that would return some
-    indication of whether it has any children, and use that to let the user know
-    that they cannot delete this particular student.  But I'm too lazy at this
-    point.
-    """
-    session.delete(student)
-
-
-def delete_student_major(sess):
-    """Undeclare a student from a particular major.
-    :param sess:    The current database session.
-    :return:        None
-    """
-    print("Prompting you for the student and the major that they no longer have.")
-    student: Student = select_student(sess)
-    major: Major = select_major(sess)
-    student.remove_major(major)
-
-
-def delete_major_student(sess):
-    """Remove a student from a particular major.
-    :param sess:    The current database session.
-    :return:        None
-    """
-    print("Prompting you for the major and the student who no longer has that major.")
-    major: Major = select_major(sess)
-    student: Student = select_student(sess)
-    major.remove_student(student)
-
-
-
-def list_student(sess: Session):
-    """
-    List all Students currently in the database.
-    :param sess:    The current connection to the database.
-    :return:
-    """
-    students: [Student] = list(sess.query(Student).order_by(Student.lastName, Student.firstName))
-    for student in students:
-        print(student)
-
-
-def list_major(sess: Session):
-    """
-    List all majors in the database.
-    :param sess:    The current connection to the database.
-    :return:
-    """
-    majors: [Major] = list(sess.query(Major).order_by(Major.departmentAbbreviation))
-    for major in majors:
-        print(major)
-
-
-def list_student_major(sess: Session):
-    """Prompt the user for the student, and then list the majors that the student has declared.
-    :param sess:    The connection to the database
-    :return:        None
-    """
-    student: Student = select_student(sess)
-    recs = sess.query(Student).join(StudentMajor, Student.studentID == StudentMajor.studentId).join(
-        Major, StudentMajor.majorName == Major.name).filter(
-        Student.studentID == student.studentID).add_columns(
-        Student.lastName, Student.firstName, Major.description, Major.name).all()
-    for stu in recs:
-        print(f"Student name: {stu.lastName}, {stu.firstName}, Major: {stu.name}, Description: {stu.description}")
-
-
-def list_major_student(sess: Session):
-    """Prompt the user for the major, then list the students who have that major declared.
-    :param sess:    The connection to the database.
-    :return:        None
-    """
-    major: Major = select_major(sess)
-    recs = sess.query(Major).join(StudentMajor, StudentMajor.majorName == Major.name).join(
-        Student, StudentMajor.studentId == Student.studentID).filter(
-        Major.name == major.name).add_columns(
-        Student.lastName, Student.firstName, Major.description, Major.name).all()
-    for stu in recs:
-        print(f"Student name: {stu.lastName}, {stu.firstName}, Major: {stu.name}, Description: {stu.description}")
-
-
-
-
-
-def select_student_from_list(session):
-    """
-    This is just a cute little use of the Menu object.  Basically, I create a
-    menu on the fly from data selected from the database, and then use the
-    menu_prompt method on Menu to display characteristic descriptive data, with
-    an index printed out with each entry, and prompt the user until they select
-    one of the Students.
-    :param session:     The connection to the database.
-    :return:            None
-    """
-    # query returns an iterator of Student objects, I want to put those into a list.  Technically,
-    # that was not necessary, I could have just iterated through the query output directly.
-    students: [Department] = list(sess.query(Department).order_by(Department.lastName, Department.firstName))
-    options: [Option] = []  # The list of menu options that we're constructing.
-    for student in students:
-        # Each time we construct an Option instance, we put the full name of the student into
-        # the "prompt" and then the student ID (albeit as a string) in as the "action".
-        options.append(Option(student.lastName + ', ' + student.firstName, student.studentId))
-    temp_menu = Menu('Student list', 'Select a student from this list', options)
-    # text_studentId is the "action" corresponding to the student that the user selected.
-    text_studentId: str = temp_menu.menu_prompt()
-    # get that student by selecting based on the int version of the student id corresponding
-    # to the student that the user selected.
-    returned_student = sess.query(Department).filter(Department.studentId == int(text_studentId)).first()
-    # this is really just to prove the point.  Ideally, we would return the student, but that
-    # will present challenges in the exec call, so I didn't bother.
-    print("Selected student: ", returned_student)
-
-
-
-
-def boilerplate(sess):
-    """
-    Add boilerplate data initially to jump start the testing.  Remember that there is no
-    checking of this data, so only run this option once from the console, or you will
-    get a uniqueness constraint violation from the database.
-    :param sess:    The session that's open.
-    :return:        None
-    """
-    department: Department = Department('CECS', 'Computer Engineering Computer Science')
-    major1: Major = Major(department, 'Computer Science', 'Fun with blinking lights')
-    major2: Major = Major(department, 'Computer Engineering', 'Much closer to the silicon')
-    student1: Student = Student('Brown', 'David', 'david.brown@gmail.com')
-    student2: Student = Student('Brown', 'Mary', 'marydenni.brown@gmail.com')
-    student3: Student = Student('Disposable', 'Bandit', 'disposable.bandit@gmail.com')
-    student1.add_major(major1)
-    student2.add_major(major1)
-    student2.add_major(major2)
-    sess.add(department)
-    sess.add(major1)
-    sess.add(major2)
-    sess.add(student1)
-    sess.add(student2)
-    sess.add(student3)
-    sess.flush()                                # Force SQLAlchemy to update the database, although not commit
-
-
-def session_rollback(sess):
-    """
-    Give the user a chance to roll back to the most recent commit point.
-    :param sess:    The connection to the database.
-    :return:        None
-    """
-    confirm_menu = Menu('main', 'Please select one of the following options:', [
-        Option("Yes, I really want to roll back this session", "sess.rollback()"),
-        Option("No, I hit this option by mistake", "pass")
-    ])
-    exec(confirm_menu.menu_prompt())
+from DatabaseManager import add, delete, list_objects, boilerplate, session_rollback
 
 
 if __name__ == '__main__':
